refactor(LazyPromise): was LazyThing
