+++ src/@littleware/little-elements/common/mutexHelper.ts
@@ -1,98 +1,3 @@
-/**
- * Little lazy loader with memory and mutex
- */
-export class LazyThing<T> implements PromiseLike<T> {
-    // tslint:disable-next-line
-    private _thing: Promise<T> = null;
-    private reload: Promise<T> = null;
-    private ttlSecs = -1;
-    // tslint:disable-next-line
-    private _lastLoadTime = 0;
-    private loader: () => T|Promise<T> = null;
-
-    /**
-     * @param loader lambda that loads the thing on demand
-     * @param ttlSecs number of seconds to cache the thing before triggering a reload,
-     *          default to never reload
-     */
-    constructor(loader: () => T|Promise<T>, ttlSecs = -1) {
-        this.loader = loader;
-        this.ttlSecs = ttlSecs;
-    }
-
-    /**
-     * Refresh if ttl expired or force true
-     * @param force
-     * @return cached this.thing and the promise
-     *    that loads the new data when reload is done
-     */
-    public refreshIfNecessary(force: boolean = false): { current: Promise<T>, next: Promise<T> } {
-        if (this._thing) {
-            // trigger reload in background if necessary
-            if (
-                null === this.reload       // not reloading already
-                && this._lastLoadTime > 0  // initial load complete
-                && (force                  // reload force requested
-                    || (this.ttlSecs > 0   // or there's a TTL and it expired
-                    && Date.now() - this._lastLoadTime > this.ttlSecs * 1000)
-                    )
-             ) {
-                this.reload = Promise.resolve(this.loader());
-                this.reload.then(
-                    () => {
-                        this._lastLoadTime = Date.now();
-                        this._thing = this.reload;
-                        this.reload = null;
-                    },
-                ).catch(
-                    (err) => {
-                        this._lastLoadTime = Date.now();
-                        this.reload = null;
-                        // leave the previously loaded thing in place
-                    },
-                );
-            }
-            return { current: this._thing, next: this.reload || this._thing };
-        }
-
-        // initial load ...
-        this._thing = Promise.resolve(this.loader());
-        this._thing.finally(() => {
-            this._lastLoadTime = Date.now();
-        });
-        return { current: this._thing, next: this._thing };
-    }
-
-    get thing(): Promise<T> {
-        return this.refreshIfNecessary(false).current;
-    }
-
-    /**
-     * Date.now of last load attempt completion -
-     * whether it succeeded or not.  Zero until after
-     * the first call to .thing
-     */
-    get lastLoadTime() {
-        return this._lastLoadTime;
-    }
-
-    /**
-     * Apply a transformation on this thing.
-     * Shortcut for new LazyThing(() => parent.thing.then(...))
-     * Note that refresh() on the child thing
-     * does not force a refresh on the parent, and
-     * vice versa, but the child inherits the ttl from
-     * its parent.
-     *
-     * @param lambda
-     */
-    public then<R>(lambda: (x: T) => R|PromiseLike<R>): LazyThing<R> {
-        const result = new LazyThing(() => {
-            return this.refreshIfNecessary(false).next.then((thing) => lambda(thing));
-        }, this.ttlSecs);
-        return result;
-    }
-}
 
 export function sleep(ms: number): Promise<void> {
     return new Promise(
